git-svn-id: http://svn.code.sf.net/p/qlandkartegt/code/QLandkarteGT/trunk@3549 d0c010d8-bb82-477e-a075-112140230451

+++ src/CTrack.h
@@ -314,9 +314,9 @@ class CTrack : public IItem
         bool highlight;
 
         /// total time covered by all track points
-        quint32 totalTime;
+        double totalTime;
         /// total time moving
-        quint32 totalTimeMoving;
+        double totalTimeMoving;
         /// total distance of track [m]
         double  totalDistance;
